app/requirements.txt
Flask==2.3.3
Pillow==10.1.0
numpy==1.26.0